# apps/knot-space/swagger.yaml
openapi: 3.0.0
info:
  title: Knot Space API
  version: 0.1.0
  description: Online repository server for Knot packages
  contact:
    name: Knot Team
    url: https://github.com/saravenpi/knot
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.knot.space
    description: Production server
...
paths:
...
  /api/packages/publish:
    post:
      tags:
        - packages
      summary: Publish a package
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                metadata:
                  type: string
                  description: JSON string of package metadata
                file:
                  type: string
                  format: binary
                  description: Package tar.gz file
      responses:
        '200':
          description: Package published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageResponse'
        '400':
          description: Invalid input
        '409':
          description: Package version already exists
        '413':
          description: File too large
...
  /api/packages/{name}:
    get:
      tags:
        - packages
      summary: Get all versions of a package
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Package versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PackageResponse'
  /api/packages/{name}/{version}:
    get:
      tags:
        - packages
      summary: Get specific package version
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Package details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageResponse'
        '404':
          description: Package not found

    delete:
      tags:
        - packages
      summary: Delete a package version
      security:
        - BearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Package deleted successfully
        '403':
          description: Forbidden
        '404':
          description: Package not found
  /api/packages/download/{packageId}:
    get:
      tags:
        - packages
      summary: Download package file
      parameters:
        - name: packageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Package file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Package not found
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
...
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        owner_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PackageResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: string
        description:
          type: string
          nullable: true
        owner:
          $ref: '#/components/schemas/UserProfile'
        team:
          $ref: '#/components/schemas/Team'
          nullable: true
        download_url:
          type: string
        file_size:
          type: integer
          format: int64
        published_at:
          type: string
          format: date-time